requirements.txt
alembic==1.15.2
altair==4.2.2
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.3
bleach==6.2.0
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cmdstanpy==1.2.5
colorama==0.4.6
coloredlogs==14.0
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dagit==1.10.10
dagster-cloud==1.10.7
dagster-cloud-cli==1.10.7
dagster-duckdb==0.26.7
dagster-graphql==1.10.10
dagster-pipes==1.10.7
dagster-shared==0.26.7
dagster-webserver==1.10.10
# Editable install with no version control (dagster_university==0.0.0)
-e /Users/user/PycharmProjects/DagsterPipeline/dagster_university
DateTime==5.5
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
duckdb==1.2.1
entrypoints==0.4
executing==2.2.0
Faker==37.1.0
fastjsonschema==2.21.1
filelock==3.18.0
flake8==7.2.0
flet==0.27.6
fonttools==4.57.0
fqdn==1.5.1
fsspec==2025.3.2
geopandas==1.0.1
github3.py==4.0.1
googleapis-common-protos==1.69.2
gql==3.5.2
graphene==3.4.3
graphql-core==3.2.4
graphql-relay==3.2.0
great-expectations==0.18.22
grpcio==1.71.0
grpcio-health-checking==1.71.0
h11==0.14.0
holidays==0.70
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
influxdb-client==1.48.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
makefun==1.15.6
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.3
monotonic==1.6
multidict==6.2.0
multipledispatch==1.0.0
narwhals==1.33.0
natsort==8.4.0
nba_api==1.9.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandas-flavor==0.6.0
pandocfilters==1.5.1
parso==0.8.4
pex==2.33.7
pexpect==4.9.0
pillow==11.1.0
pip_search==0.0.13
platformdirs==4.3.7
pluggy==1.5.0
polars==1.27.1
posthog==3.23.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
prophet==1.1.6
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pyflakes==3.3.2
Pygments==2.19.1
pyjanitor==0.31.0
PyJWT==2.10.1
pymongo==4.11.3
pyogrio==0.10.0
pyparsing==3.2.3
pyproj==3.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
questionary==2.1.0
reactivex==4.0.4
referencing==0.36.2
repath==0.9.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==78.1.0
sgmllib3k==1.0.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
stanio==0.5.1
starlette==0.46.1
structlog==25.2.0
suds-community==1.2.0
suds-py3==1.4.5.0
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.2.1
toolz==1.0.0
toposort==1.10
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
universal_pathlib==0.2.6
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchdog==5.0.3
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.13
wrapt==1.17.2
xarray==2025.3.1
xgboost==3.0.0
yarl==1.18.3
zipp==3.21.0
zope.interface==7.2
